feat: 메인페이지 경로 변경
